+++ src/core/maps/store/actions/actionsMap.js
@@ -117,8 +117,30 @@ export default {
      * @returns {OLCesium} - ol cesium map.
      */
     createMap3D ({getters}) {
+        const backwardsConfigCesiumParameter = {...Config?.cesiumParameter};
+
+        /**
+         * @deprecated in the next major-release!
+         * Backward compatibility: globe parameters are set in the object globe now
+         */
+        if (!backwardsConfigCesiumParameter.globe) {
+            backwardsConfigCesiumParameter.globe = {};
+        }
+        if (backwardsConfigCesiumParameter.enableLighting) {
+            backwardsConfigCesiumParameter.globe.enableLighting = backwardsConfigCesiumParameter.enableLighting;
+            console.warn("The attribute 'cesiumParameter.enableLighting' is deprecated. Please use 'cesiumParameter.globe.enableLighting'!");
+        }
+        if (backwardsConfigCesiumParameter.maximumScreenSpaceError) {
+            backwardsConfigCesiumParameter.globe.maximumScreenSpaceError = backwardsConfigCesiumParameter.maximumScreenSpaceError;
+            console.warn("The attribute 'cesiumParameter.maximumScreenSpaceError' is deprecated. Please use 'cesiumParameter.globe.maximumScreenSpaceError'!");
+        }
+        if (backwardsConfigCesiumParameter.tileCacheSize) {
+            backwardsConfigCesiumParameter.globe.tileCacheSize = backwardsConfigCesiumParameter.tileCacheSize;
+            console.warn("The attribute 'cesiumParameter.tileCacheSize' is deprecated. Please use 'cesiumParameter.globe.tileCacheSize'!");
+        }
+
         return api.map.createMap({
-            cesiumParameter: Config?.cesiumParameter,
+            cesiumParameter: backwardsConfigCesiumParameter,
             map2D: getters.get2DMap,
             shadowTime: function () {
                 return this.time || Cesium.JulianDate.fromDate(new Date());